Some more fixes

+++ src/Result.kt
@@ -1,1 +1,1 @@
-data class Result(val time: Double = Double.MAX_VALUE, val optimizations: List<String> = emptyList())
+data class Result(val time: Int = Int.MAX_VALUE, val optimizations: List<String> = emptyList())
